pyproject.toml
[project]
name = "gaj-ai-scientist"
version = "0.1.0"
description = ""
authors = [
    {name = "Gajesh Naik",email = "robogajesh@gmail.com"}
]
readme = "README.md"
requires-python = "<4.0,>=3.11"
dependencies = [
    "openai (>=1.84.0,<2.0.0)",
    "anthropic (>=0.52.2,<0.53.0)",
    "instructor (>=1.8.3,<2.0.0)",
    "requests (>=2.31.0,<3.0.0)",
    "semanticscholar (>=0.7.0,<1.0.0)",
    "arxiv (>=2.2.0,<3.0.0)",
    "pdf2image (>=1.17.0,<2.0.0)",
    "pillow (>=10.0.0,<11.0.0)",
    "rich (>=13.0.0,<14.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
